# website/pages/assets/openapi.yaml
openapi: 3.1.0
info:
  title: PicoMQ HTTP API
  version: 0.1.0
  description: >
    The Pico protocol, PicoMQ's native HTTP API, plus the admin API served by
    each node's admin listener. A stream is addressed by its path. Requests for
    a stream owned by another node return 307 with the owner's address. With
    auth required, send `Authorization: Bearer <token>` (probes stay open).
servers:
  - url: http://127.0.0.1:4437
    description: Protocol listener
...
security:
  - BearerAuth: []
paths:
...
  /_schemas/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: The schema name. Names are paths and may contain slashes.
        schema: { type: string }
    put:
      tags: [schemas]
      summary: Register a schema
      description: >
        Store or replace a schema. The format comes from the Content-Type, or
        from an extension on the name. 400 when the schema does not parse or
        the node runs without a schema registry.
      operationId: putSchema
      requestBody:
        required: true
        content:
          application/schema+json:
            schema: { type: string }
          application/avro:
            schema: { type: string }
          application/x-protobuf:
            schema: { type: string }
      responses:
        "204":
          description: Stored.
        "400":
          description: Unparseable schema, unknown format, or no registry configured.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [schemas]
      summary: Fetch a schema
      operationId: getSchema
      responses:
        "200":
          description: The schema body with its content type.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such schema.
    delete:
      tags: [schemas]
      summary: Delete a schema
      operationId: deleteSchema
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such schema.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Unauthenticated:
      description: Missing or invalid credential.
      headers:
        WWW-Authenticate:
          schema: { type: string, enum: [Bearer] }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Valid credential, insufficient scope (or producer fencing on append).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        next_seq: { type: integer }
...